backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.5.2
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
easydict==1.9
editdistance==0.8.1
efficientnet==1.0.0
essential-generators==1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
idna==3.6
imageio==2.34.1
imgaug==0.4.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jax==0.4.25
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.10.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.5.0
lazy_loader==0.4
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.2.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.10
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchmetrics==0.2.0
torchvision==0.17.2+cu118
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
